# enterprise/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Boundless Enterprise Multipass API
  description: |
    Enterprise-grade unified access and control layer for the Boundless blockchain platform.

    ## Features
    - Post-Quantum Cryptography (Dilithium5 + Kyber1024)
    - KYC/AML Identity Management
    - Multi-Asset Wallet System
    - Internal Asset Trading Market
    - Hardware Pass NFC Integration
    - SSO & Session Management
    - Enterprise Application Registry
    - Events & Reporting

    ## Security
    - All endpoints (except `/api/auth/login` and `/api/auth/register`) require authentication
    - JWT-based session tokens
    - Rate limiting on sensitive endpoints
    - Encrypted private key storage with master encryption key

    ## Base URL
    `http://localhost:8080` (default development)

  version: 1.0.0
  contact:
    name: Boundless Platform
    url: https://boundless.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.boundless.example.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/api/auth/login`
  schemas:
...
    Identity:
      type: object
      properties:
        identity_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        kyc_status:
          type: string
          enum: [Pending, Verified, Rejected, Expired]
        created_at:
          type: string
          format: date-time
    Wallet:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        identity_id:
          type: string
          format: uuid
        wallet_type:
          type: string
          enum: [Personal, Enterprise, Escrow]
        boundless_address:
          type: string
          description: 64-character hex blockchain address
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/identity/create:
    post:
      tags: [Identity]
      summary: Create identity profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, full_name]
              properties:
                email:
                  type: string
                  format: email
                full_name:
                  type: string
                phone:
                  type: string
                country:
                  type: string
      responses:
        '200':
          description: Identity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
  /api/identity/{id}:
    get:
      tags: [Identity]
      summary: Get identity by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Identity profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
  /api/identity/{id}/kyc-status:
    put:
      tags: [Identity]
      summary: Update KYC status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [Pending, Verified, Rejected, Expired]
      responses:
        '200':
          description: Status updated

  # =============================================================================
  # WALLET
  # =============================================================================
  /api/wallet/create:
    post:
      tags: [Wallet]
      summary: Create blockchain wallet
      description: Generates PQC keypair and derives Boundless address
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity_id, wallet_type]
              properties:
                identity_id:
                  type: string
                  format: uuid
                wallet_type:
                  type: string
                  enum: [Personal, Enterprise, Escrow]
      responses:
        '200':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
  /api/wallet/{id}/balances:
    get:
      tags: [Wallet]
      summary: Get wallet balances
      description: Returns blockchain balances for all assets
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallet_id:
                    type: string
                    format: uuid
                  balances:
                    type: array
                    items:
                      type: object
                      properties:
                        asset:
                          type: string
                        amount:
                          type: integer
                          format: int64
  /api/wallet/{id}/transfer:
    post:
      tags: [Wallet]
      summary: Transfer blockchain assets
      description: Creates and submits signed transaction to Boundless blockchain
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_address, amount]
              properties:
                to_address:
                  type: string
                  description: 64-character hex Boundless address
                amount:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Transfer successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx_hash:
                    type: string
                    description: Blockchain transaction hash

  # =============================================================================
  # ASSETS
  # =============================================================================
...
security:
  - BearerAuth: []